+++ engine/src/main/java/com/jsoftware/platform/repository/MonthlyPickRepositoryImpl.java
@@ -28,7 +28,7 @@ public class MonthlyPickRepositoryImpl implements MonthlyPickRepository{
     }
 
     @Override
-    public MonthlyPick readMonthlyPick(Long id) {
+    public MonthlyPick readMonthlyPick(int id) {
         return sqlSessionTemplate.selectOne("selectMonthlyPickById", id);
     }
