+++ engine/src/main/java/com/jsoftware/platform/repository/RecommendActivityRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface RecommendActivityRepository {
     List<RecommendActivity> readRecommendActivities();
 
-    RecommendActivity readRecommendActivity(Long id);
+    RecommendActivity readRecommendActivity(int id);
 
     void createRecommendActivity(RecommendActivity recommendActivity);
 }
