+++ engine/src/main/java/com/jsoftware/platform/repository/TodaysHeadlineRepositoryImpl.java
@@ -27,7 +27,7 @@ public class TodaysHeadlineRepositoryImpl implements TodaysHeadlineRepository{
     }
 
     @Override
-    public TodaysHeadline readTodaysHeadline(Long id) {
+    public TodaysHeadline readTodaysHeadline(int id) {
         return sqlSessionTemplate.selectOne("selectTodaysHeadlineById", id);
     }
