+++ engine/src/main/java/com/jsoftware/platform/service/RecommendHotspotService.java
@@ -23,7 +23,7 @@ public class RecommendHotspotService {
     }
 
     @Transactional
-    public RecommendHotspot readRecommendHotspot(Long id) {
+    public RecommendHotspot readRecommendHotspot(int id) {
         return repository.readRecommendHotspot(id);
     }
